Calculating and displaying right probability on result activity.

+++ app/src/main/java/com/robertoallende/diagnosis/model/DiagnosisPlan.java
@@ -26,7 +26,7 @@ public class DiagnosisPlan {
         plan.saveUnansweredAnswer(binaryAnswer);
     }
 
-    public int getProbability() {
+    public float getProbability() {
         return plan.getProbability();
     }
 }
